+++ Sprint1/src/app/app-routing.module.ts
@@ -1,26 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AdminComponent } from './admin/admin.component';
-import { LoginComponent } from './login/login.component';
-import { NewsfeedComponent } from './newsfeed/newsfeed.component';
-import { ProfileComponent } from './profile/profile.component';
-import { SettingComponent } from './setting/setting.component';
-import { RegisterComponent } from './login/register/register.component';
-import { HomeComponent } from './home/home.component';
+import { AdminComponent } from './module/admin/admin.component';
+import { NewsfeedComponent } from './module/newsfeed/newsfeed.component';
+import { ProfileComponent } from './module/profile/profile.component';
+import { SettingComponent } from './module/setting/setting.component';
+import { LoginComponent } from './component/login/login.component';
 
 const routes: Routes = [
   // { path: '', redirectTo: '/login', pathMatch: 'full' },
   {
     path: 'login',
-    component: HomeComponent,
-  },
-  {
-    path: 'login2',
     component: LoginComponent,
-  },
-  {
-    path: 'register',
-    component: RegisterComponent,
   },
   {
     path: 'admin',
