+++ Sprint1/src/app/component/likelist/likelist.component.ts
