+++ Sprint1/src/app/component/register/register.component.html
@@ -1,0 +1,3 @@
+<p>
+  register works!
+</p>
