+++ Sprint1/src/app/component/userinfo/userinfo.component.scss
