--- Sprint1/src/app/login/login.component.html
@@ -1,34 +1,0 @@
-<div class="row mt-5">
-  <div class="col-md-6 m-auto">
-    <div class="card text-dark bg-light mb-3">
-      <h1 class="text-center mt-3">Login</h1>
-
-      <form action="" method="POST">
-        <div class="mb-3 p-3">
-          <label for="email" class="form-label">Email</label>
-          <input
-            type="email"
-            id="email"
-            name="email"
-            class="form-control"
-            placeholder="Enter Email"
-          />
-        </div>
-        <div class="mb-3 p-3">
-          <label for="password" class="form-label">Password</label>
-          <input
-            type="password"
-            id="password"
-            name="password"
-            class="form-control"
-            placeholder="Enter Password"
-          />
-        </div>
-        <button type="submit" class="btn btn-primary btn-block mx-3">
-          Login
-        </button>
-      </form>
-      <p class="lead mt-4 mx-3">No Account? <a href="/register">Register</a></p>
-    </div>
-  </div>
-</div>
