--- Sprint1/src/app/login/login.module.ts
@@ -1,9 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { LoginComponent } from './login.component';
-
-@NgModule({
-  declarations: [LoginComponent],
-  imports: [CommonModule],
-})
-export class LoginModule {}
