--- Sprint1/src/app/login/register/register.component.html
@@ -1,56 +1,0 @@
-<div class="row mt-5">
-  <div class="col-md-6 m-auto">
-    <div class="card text-dark bg-light mb-3">
-      <h1 class="text-center mt-3">Register</h1>
-
-      <form action="" method="POST">
-        <div class="mb-2 p-3">
-          <label for="name" class="form-label">Name</label>
-          <input
-            type="text"
-            id="name"
-            name="name"
-            class="form-control"
-            placeholder="Enter Name"
-          />
-        </div>
-        <div class="mb-2 p-3">
-          <label for="email" class="form-label">Email</label>
-          <input
-            type="email"
-            id="email"
-            name="email"
-            class="form-control"
-            placeholder="Enter Email"
-          />
-        </div>
-        <div class="mb-2 p-3">
-          <label for="password" class="form-label">Password</label>
-          <input
-            type="password"
-            id="password"
-            name="password"
-            class="form-control"
-            placeholder="Enter Password"
-          />
-        </div>
-        <div class="mb-2 p-3">
-          <label for="confirmPassword" class="form-label"
-            >Confirm Password</label
-          >
-          <input
-            type="password"
-            id="confirmPassword"
-            name="confirmPassword"
-            class="form-control"
-            placeholder="Confirm Password"
-          />
-        </div>
-        <button type="submit" class="btn btn-primary btn-block mx-3">
-          Register
-        </button>
-      </form>
-      <p class="lead mt-4 mx-3">Have An Account? <a href="/login">Login</a></p>
-    </div>
-  </div>
-</div>
