+++ Sprint1/src/app/module/admin/admin.component.scss
