+++ Sprint1/src/app/module/newsfeed/newsfeed.component.html
