+++ Sprint1/src/app/module/newsfeed/newsfeed.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NewsfeedComponent } from './newsfeed.component';
-import { LikelistComponent } from './likelist/likelist.component';
+import { LikelistComponent } from '../../component/likelist/likelist.component';
 import {MatIconModule} from '@angular/material/icon'
 
 @NgModule({
