+++ Sprint1/src/app/module/setting/setting.component.scss
