+++ Sprint1/src/app/services/userlist.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { User } from '../admin/interface/user.model';
+import { User } from 'src/app/services/interface/user.model';
 
 @Injectable({
   providedIn: 'root',
