+++ Sprint1/src/styles.scss
@@ -4,15 +4,7 @@
     padding: 0;
     box-sizing: border-box;
 }
-html{
+html, body, app-root{
     height: 100%;
 }
-body {
-    height: 100%;
-
-}
-app-root {
-    height: 100%;
-}
-html, body { height: 100%; }
 body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
